requirements.txt
mysql-connector-python==8.0.33
tkinter
Pillow==10.0.0 